requirements.txt
fastapi
uvicorn[standard]
Pillow
numpy
torch
torchvision
selectivesearch
opencv-contrib-python
python-multipart
aiofiles
